+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 
 ### Added
 
-* New instances to `dx run --instance-type-help` to correspond [official docs](https://documentation.dnanexus.com/developer/api/running-analyses/instance-types)
+* New instance types in `dx run --instance-type-help` corresponding to the [official docs](https://documentation.dnanexus.com/developer/api/running-analyses/instance-types)
 
 ## [398.0] - beta